+++ QSB/QSB.csproj
@@ -94,9 +94,9 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="AnimationSync\AnimationSync.cs" />
-    <Compile Include="AnimationSync\AnimatorMirror.cs" />
-    <Compile Include="AnimationSync\AnimFloatParam.cs" />
+    <Compile Include="Animation\AnimationSync.cs" />
+    <Compile Include="Animation\AnimatorMirror.cs" />
+    <Compile Include="Animation\AnimFloatParam.cs" />
     <Compile Include="DebugLog.cs" />
     <Compile Include="Messaging\MessageHandler.cs" />
     <Compile Include="Messaging\MessageType.cs" />
